lib.c: fix implementation of File-Access word set routines
Two main problems:

1. fid’s were generated starting at 0. The standard does not explicitly
forbid this, but 0 is used in SOURCE-ID to mean the terminal, so it
doesn’t work.

2. FILE-POSITION and REPOSITION-FILE used single-length counts, where
the standard specifies doubles.

Additionally, no attempt was made to check that the file IDs passed in
were valid (now checked). Also increase the number of file descriptors
available to 64.

Doing this properly, by passing through FILE * pointers, would require
allowing Beetle to match its cell to the native word length.

Also add a FIXME to define and use CHECKA, as currently only the defined
LIB codes even call it, and it’s undefined.