step.c: change how address checking is done
Make CHECKC and CHECKA check Beetle addresses rather than checking real
addresses directly; have CHECKP to check a pointer that is assumed to be
a pointer to Beetle’s main memory.

Most calls to CHECKA become calls to CHECKP, except those in the fetch and
store instructions (@ ! +!).

Later, we will redefine CHECKP to be indepedent of the other macros, which
will themselves be allowed to check other memory mappings.