Add himem implementation, with tests
See beetle.h for the new himem_* APIs. These are particular to C Beetle.

With them, C memory can be mapped directly to Beetle memory.

Currently the memory allocation model works like Forth’s HERE and ALLOT: it
is only possible to allocate Beetle address space, not free it. Allocations
are contiguous; but the “here” pointer can be cell-aligned.

Currently the maximum number of allocations is also hardwired, to 256.