Make M0 a CELL *, not a BYTE *
This fixes a bunch of pointer alignment warnings (thanks, Clang on
Android!).

Also find and fix a couple of bugs where cell counts were being used instead
of byte counts or vice versa (although not betting against having introduced
more with this wide-ranging patch).

Replace some instances of clearing to zero with memset or calloc.

As well as the tests passing, pForth correctly builds itself with the new
Beetle.