step.c: add address checks for argument to SP! and RP!
This changed the results of an exceptst test, so update the expected result.

This meant that no exceptst test actually tested invalid SP during address
error (invadr/aliadr labels, code -258), so add one for that. Also add one
for bad 'THROW (code -259).

Remove a previous test for code -259 that had no expected result and was
never executed (old test no. 13).

Change the loop that executes the tests to replace a constant number of
tests with a calculation based on the size of the expected results array, so
that all tests are guaranteed to be attempted.