Check arguments to init_beetle are sensible
Check c_array is non-NULL, and size is at least 4.

If not, return -1 (as per spec), not 1 (as code did previously).

Move remaining tests of init_beetle into initialt.c from stept.c. (This may
have made sense in the past when initialt tested compilation, but that job
is now handled by autoconf.)