Restore MEMORY register
Until we have a way for bForth programs to read the memory map, and
position-independent code (at least for the build of pForth), allow them
at least to get main memory, and guarantee that it starts at address 0.

Undo corresponding changes in tests.

Also map 'BAD and -ADDRESS read-only into main memory, and remove some
direct memory accesses that prevented that working before.

This completes the memory-mapping of registers.