Split parts of beetle.h into beetle_aux.h and beetle_private.h
beetle_aux.h is an undocumented public header. It is useful but not strictly
necessary for clients of Beetle. The UI and tests use it.

beetle_private.h is only used internally.

Also don’t use the gnulib intprops module, which hasn’t been needed for some
time (it was used for 2’s complement checks, which are now assumed by the
ISO C99 definition).