Remove memory mapping, restore M0, R0, S0; remove #S, #R
Memory mapping was an interesting experiment, but it’s not very “Beetle”,
and the implementation was quite inefficient.

I want a “final” version of Beetle that is true to its genesis, and also as
a fair comparison for Mit (Beetle is already rather faster at running
pForth, but should be faster still!).