Fix specification and implementation of (+LOOP)[I] (fix #9)
@apt1002 pointed out that the specification of (+LOOP)[I] was ambiguous: the
loop increment must be a signed number (as it is in the ANSI and 2012 Forth
standards).

Further, @apt1002 pointed out that the implementation was buggy: it
considered one bit too few. The natural way is to check overflow, but
failing that, there’s a neat fix in GForth:

http://git.savannah.gnu.org/cgit/gforth.git/tree/prim#n373

Apply this fix to Beetle. Many thanks, @apt1002!