Use “inline” directly
This avoids a bug in gnulib’s _GL_EXTERN_INLINE macro at the cost of
assuming a working C99 compiler; but we already do that.

And in fact there’s no need to use “extern inline”: if we keep the public
declaration of the function unchanged (implicit “extern”, no “inline”) then
the function is available externally, and inlined in step.c.